do not inherit from STL
