+++ src/openms/source/KERNEL/ConsensusMap.cpp
@@ -50,9 +50,9 @@ namespace OpenMS
     MetaInfoInterface(),
     RangeManagerContainerType(),
     DocumentIdentifier(),
+    ExposedVector<ConsensusFeature>(),
     UniqueIdInterface(),
     UniqueIdIndexer<ConsensusMap>(),
-    data_(),
     column_description_(),
     experiment_type_("label-free"),
     protein_identifications_(),
@@ -65,9 +65,9 @@ namespace OpenMS
     MetaInfoInterface(source),
     RangeManagerContainerType(source),
     DocumentIdentifier(source),
+    ExposedVector<ConsensusFeature>(source),
     UniqueIdInterface(source),
     UniqueIdIndexer<ConsensusMap>(source),
-    data_(source.data_),
     column_description_(source.column_description_),
     experiment_type_(source.experiment_type_),
     protein_identifications_(source.protein_identifications_),
@@ -81,9 +81,8 @@ namespace OpenMS
   ConsensusMap::ConsensusMap(Base::size_type n) :
     MetaInfoInterface(),
     RangeManagerContainerType(),
-    DocumentIdentifier(),
+    ExposedVector<ConsensusFeature>(n),
     UniqueIdInterface(),
-    data_(n),
     column_description_(),
     experiment_type_("label-free"),
     protein_identifications_(),
